Update lomb_scargle.py and delete new file
